# Implement HTM without eval