As pointed out in #51, the current version of HTM uses code evaluation. This prohibits setting a strict Content-Security Policy.

This pull request implements HTM *without* code generation. Instead the parsing phase generates a list of "instructions", which the result can be evaluated from. For each instruction we append to the list FIELD, VALUE and OP, and for one special instruction (PROPS_SET) we append an additional KEY after the OP. Possible OPs are:
 * TAG_SET: set the tag to `FIELD ? fields[FIELD] : VALUE`
 * PROPS_SET: set prop `KEY` to `FIELD ? fields[FIELD] : VALUE`
 * PROPS_ASSIGN: Object.assign `FIELD ? fields[FIELD] : VALUE` to props
 * CHILD_RECURSE: evaluate `FIELD ? fields[FIELD] : VALUE` and append to children
 * CHILD_APPEND: append `FIELD ? fields[FIELD] : VALUE` to children

For some of these OPs for example VALUE is not relevant, but they're there anyway (set to `0`) to keep the evaluation code compact.

To compress the parsing logic `MODE_SKIP` is now replaced by `MODE_SLASH`, signifying "we're still parsing a tag, but have encountered a slash, so ignore everything until a TAG_END".

The .gz/.br code size is smaller compared to the current master branch version. The "creation" benchmark numbers are higher, but the "usage" benchmark is lower. Note that the benchmarks were run without Jest.

**Before change**

```
Build "htm" to dist:
        649 B: htm.mjs.gz
        596 B: htm.mjs.br
        714 B: htm.umd.js.gz
        645 B: htm.umd.js.br
Creation: 37,783/s, average: 26µs
Usage: 520,896/s, average: 1µs
```

**After change**

```
Build "htm" to dist:
        602 B: htm.mjs.gz
        544 B: htm.mjs.br
        669 B: htm.umd.js.gz
        594 B: htm.umd.js.br
Creation: 43,598/s, average: 22µs
Usage: 364,405/s, average: 2µs
```

The Babel plugin needed to be modified, as the props assignment pattern changed.