@jviide 2018-12-27
Pushed three suggested commits that can be rolled back if needed:

 * 6752939 implements the support for multiple root elements.
 * d060bb8 compresses code size by using String#indexOf. Based on [this StackOverflow answer](https://stackoverflow.com/a/21772672) IE6 supports it for strings.
 * 7199d5c compresses code size by using property access instead of String#charAt.  Based on the [ECMAScript 5](https://kangax.github.io/compat-table/es5/) compatibility IE8 supports it.

These changes together make the code size smaller, but with some performance loss (mostly from multi-root support, creation benchmark still being faster than the master version):
```
Build "htm" to dist:
        579 B: htm.mjs.gz
        521 B: htm.mjs.br
        645 B: htm.umd.js.gz
        581 B: htm.umd.js.br
Creation: 39,616/s, average: 25µs
Usage: 344,077/s, average: 2µs
```