@jviide 2018-12-27
It should be mentioned that at least one @mikesherov's size optimization from #48 is also included in this PR.