@mikesherov 2018-12-27
Ha, beat me to the punch. I was literally fiddling around with getting rid of charAt and charCodeAt myself.