@jviide 2018-12-27
Rolling back the multi-root support but keeping everything else in this PR gives the following numbers:
```
Build "htm" to dist:
        560 B: htm.mjs.gz
        500 B: htm.mjs.br
        624 B: htm.umd.js.gz
        554 B: htm.umd.js.br
Creation: 42,794/s, average: 23µs
Usage: 357,357/s, average: 2µs
```