@jviide 2019-01-16
After the latest code changes the code sizes are:
```
Build "htm" to dist:
        539 B: htm.mjs.gz
        488 B: htm.mjs.br
        607 B: htm.umd.js.gz
        545 B: htm.umd.js.br
```
And the perf-test-harness benchmarks:
```
  performance
    ✅  creation
       Creation: 81,423/s, average: 12µs
    ✅  usage
       Usage: 404,964/s, average: 2µs
```