@jviide 2019-01-21
The latest version is one we code-golfed together with @developit. It uses Map for caching when available, and otherwise falls back to current the string key based caching scheme. This boosts performance quite a bit on platforms where a native Map implementation is available, with no performance loss on platforms that don't. The current benchmark numbers on a platform that supports Map natively are:
```
Creation: 83,069/s, average: 12µs
Usage: 717,657/s, average: 1µs
```
The corresponding numbers on the current master branch are:
```
Creation: 15,978/s, average: 62µs
Usage: 471,498/s, average: 2µs
```

The file sizes got bigger, but mostly stayed under 600 bytes:
```
Build "htm" to dist:
        591 B: htm.mjs.gz
        540 B: htm.mjs.br
        654 B: htm.umd.js.gz
        584 B: htm.umd.js.b
```