@jviide 2019-02-20
Closing this in favor of #62, which essentially includes the features of this PR as well as #59.